Set max-line-length in flake8 to 80 (really to 88)
Now that the configuration file is fixed, max-line-length set to 80 works correctly. Black's default is 88, but the E501 check in flake8 is turned off in favour of the B950 check from flake8-bugbear, which allows lines to exceed the limit by 10%, so setting max-line-length to 80 in reality sets it to 88.